Allows levels to be added
To add levels to the game, go to multiplayer.gd, add your scene to the maptext variable array, then add that to the maps variable in a preload function. Next open the Map Selected script and add a StageButton node. Make sure to change the instance variables of the scene and imgFile to your scene directory and your image directory. Finally, go to your scene script and change the Global.update_respawn() to your preferred location. This part may require trial and error.